refactoring starter project
