+++ Examples/starter/src/main/java/com/application/project/DemoApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.application.project;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
