+++ Examples/starter/src/main/java/com/application/project/myapi/Flymsg.java
@@ -1,0 +1,13 @@
+package com.application.project.myapi;
+
+import com.flyspring.autoroute.FlyRequest;
+import com.flyspring.autoroute.annotations.PathVariableAnnotation;
+
+public class Flymsg {
+
+    @PathVariableAnnotation(name = { "{msg}" })
+    public String flyget(FlyRequest request){
+        return request.getPathVariable("msg");
+    }
+    
+}
